Remove cpuScaled as the cpu-usage metric is scaled to 100%
